code/requirements.txt
beautifulsoup4==4.11.1
mysql_connector_repackaged==0.3.1
numpy==1.22.4
pandas==1.4.2
requests==2.28.1
